Merge reflow with changeset pulses queued for the same node.
This is a major change to the Graph scheduler and collector nodes.
Previously, collectors expected pulses to be either a changeset or
reflow. The Graph scheduler queued non-reflow pulses ahead of reflow
ones, and pruned multiple reflows of the same branch by comparing
timestamps. In some cases (similar to uwdata/vega#76), this could result
in mistaken skips of reflows. The new scheduler merges reflow and
changeset pulses queued for the same node, with collectors responding
similarly. To ensure this can be done safely, we verify that there is at
most one changeset pulse for every queued node. This change should also
increase the expressivity of signal+streaming data APIs, allowing for
both to occur within the same propagation cycle.