It seems the shortcut check should be with angles, not radius
With many (100,000) circle symbols, the shortcut was not being taken
resulting in over 30% of the time in the update() for arc calculating
sin and cos. This fix lowers the time evaluating arc bounds to about 5%
and increases the frame rate.

I think this check is sufficient for the circle symbol, but there may be
able to be a stronger check for other full-circle arcs (other multiples of tau,
negative multiples of tau?).